Add necessary HTTomo params to wrapper function params
The type of parameters that would possibly be needed by tomopy wrapper
functions are not limited to loader output (ie, data), but can include
things like an MPI communicator object.

To handle this, a more generic function has been added which can be
given a list of potential HTTomo-related parameters that a tomopy
wrapper function may need, and add in the necessary parameter values
based on the wrapper function's type signature.