Fix normalisation processing in `simple_tomopy_pipeline.yaml`
The use of the `tomopy.prep.normalize.minus_log()` function was missing
in `simple_tomopy_pipeline.yaml` and thus causing the result to be
incorrect.

Adding in the use of this method revealed the need to have some logic to
determine if a wrapped method in tomopy does or doesn't need the darks
and flats as parameters.