Generically define data param key in method param dict
Up until now, the parameter in a method function which represents the
input array has been assumed to be called `data`. This is the case for
all existing tomopy wrapper functions, but may not be the case for all
HTTomo method functions (at least, it can't be assumed until a
standardisation of certain method function parameters is defined).

A function has been added to the runner which will take into account the
possibility that an HTTomo method function could have its input data
parameter be named something other than `data` (which, if not taken into
account, could lead to the parameter dict unpacking to fail when running
HTTomo method functions due to an incorrect parameter name being passed
for the input array).