Include `data_in_multi` and `data_out_multi` params in dataset initialisation
The method parameters `data_in_multi` and `data_out_multi` have been
introduced. They allow a method to be applied to multiple input datasets
and produce multiple output datasets respectively. From the introduction
of these parameters, it is necessary to check their values in the user
config YAML when initialising the dict of all datasets in the task
runner.