Enable methods to be applied to multiple inputs
Also, include the name of the output dataset in the filename of any
intermediate file. This isn't strictly necessary for when the method is
only applied to one input dataset, but needs to be specified when the
method is applied to multiple input datasets to distinguish between the
different input datasets that were processed by the single method.