Pass global stats to methods that require them
For now, the input of statistics to methods will be hidden from the
user. However, in the future it is quite possible that some
statistics-related parameters of a method will be exposed through the
user config YAML in some manner.

Note that `httomo.misc.images.save_to_images` doesn't yet work with the
runner due to an error regarding the absence of the `data_out`
parameter. So even though the functionality is now in place to pass the
necessary statistics to the image saver, that method isn't yet working
with the runner.