Workaround to reslice recon input (fix for MPI runs)
The absence of reslicing the input data of the reconstruction method to
be in the sinogram dimension rather than the projection dimension was
causing the reconstruction to be incorrect when running with more than
one MPI process.

The change here should be considered a workaround rather than a proper
solution, since more generically handling the reslicing of the input
data of methods will be necessary in the future.