Refactor functions getting darks/flats
There is now a helper function `_get_darks_flats()` which contains the
duplicated code between what used to be called `get_darks_flats()` and
`_get_separate_darks_flats()` (now called `get_darks_flats_together()`
and `get_darks_flats_separate()` respectively).

With this refactoring, if the darks/flats are in the same dataset within
the same NeXuS file, then the NeXuS file is opened once only to load
both the darks and flats.

If the darks/flats are in separate files, then there will be two
instances of opening a file.

However, if the darks/flats are in the same file but in different
datasets within that file, then the file will be opened twice instead of
just once, due to calling `get_darks_flats_separate()` twice. It may be
possible to avoid unnecessarily opening and closing the same file twice
for this particular case, in which case it'll hopefully be done in the
future.