Add classes for YAML loader to recognise param sweeps via YAML tags
This is experimenting with using a YAML "local tag" to mark in a
pipeline where a parameter sweep is being defined.

There are two ways a parameter sweep can be defined:
- explicitly enumerating all values to sweep over
- defining a range via `start`, `stop`, `step` values

Since it is conceivable that a method parameter value could be a list, a
list of values wouldn't suffice to unambiguously define a parameter
sweep (ie, a list of parameter values to sweep over, and a parameter
which took a list as its value, would be indistinguishable).

The proposed strategy is to use tuples as the data type to define a
parameter sweep, and therefore as an indicator to the httomo task runner
when it has encountered a parameter sweep.

In this change, `Sweep` is simply creating a `tuple` out of the `list`
of explicit values it is given. `SweepRange` is a bit more complicated,
generating a numpy array of values from the `start`, `stop`, `step` info
in the `dict` that it receives, before then converting that to a
`tuple`.