Move method execution code into separate function
Prior to this change, there was a large `if`/`else` block in the main
loop of the task runner, which did different things if a loader function
was being run (the `if` block) or a method function was being run (the
`else` block).

The code in the `else` block handling the execution of a method function
has been moved to a separate function.