Implement parameter sweep for final method in pipeline
There are multiple cases for which parameter sweeps need to be handled,
the simplest one being when the method with the parameter sweep is the
last one in the pipeline.

The result of each individual value in the parameter sweeps is saved as
a separate dataset within a single hdf5 file. For example, a parameter
sweep containing 5 values will result in a single hdf5 file with 5
datasets within it.

Other cases that need to be handled which are not yet implemented are:
- when another method runs after the method with the parameter sweep
  (ie, the method with the parameter sweep is not the last method to run
  in the pipeline)
- when the method with the parameter sweep runs prior to a reslice
  operation