Store global stats in a list
In the case of parameter sweeps, there are multiple arrays resulting
from a method. If these multiple arrays then get passed to some other
method requiring global stats (such as `save_to_images()`), the global
stats will need to be stored in such a way that allows associating stats
to each of the arrays resulting from the parameter sweep.

This change makes the values of the dataset-name keys in the dicts in
`glob_stats` be lists rather than single tuples, to pave the way to
get the above functionality in future changes.