Enable datasets loop to handle a method with multiple output datasets and one input dataset
Before, a constraint was placed on the number of input and output
datasets for a particular method such that they had to be the same. More
explicitly:
- if there was 1 input dataset, there had to be 1 output dataset,
  defined with `data_in` and `data_out` respectively
- if there were `n` input datasets, there had to be `n` output datasets,
  defined with `data_in_multi` and `data_out_multi` respectively

This made it difficult to handle a method which had only 1 input
dataset, but multiple output datasets.

The loop over the datasets is less constraining now, and it is possible
to define a method with 1 input dataset (with `data_in`) and multiple
output datasets (with `data_out`). See the `find_center_360` method
configuration in
`samples/pipeline_template_examples/DLS/02_i12_360scan_pipeline.yaml`
for an example.