Enable fetching of specific entry in methods database
There are plans to store more information about methods in addition to
the pattern in the "methods databse" YAML files, such as the padding
that a method will apply to its input data.

With this in mind, the ability to get a single piece of information
about a method that is stored in the "methods databse" YAML files has
been added (as opposed to having to fetch the entire dict).