Fix incorrect gathering of data when saving intermediate dataset
See issue #67 for more information on the fixed bug.

Note that a reconstruction method is a special case in determining which
dimension is the slicing dimension (see the issue on github for more
info), and thus needs to be handled differently to other methods.

The code was already using the existence of the `algorithm` parameter to
detect if a reconstruction method was being used (in order to pass the
reconstruction algorithm name to the function saving the intermediate
file), so that same information was used to catch a reconstruction method,
and then define `slice_dim` accordingly in that special case.