Modify global `run_out_dir` var in all MPI processes
Even though `httomo.globals.run_out_dir` is a global variable, each MPI
process will have its own copy of this global variable. Therefore, if
only one MPI process modifies the value from the initial value of
`None`, the other MPI processes will not see this and will still see
`None` for their copy of `httomo.globals.run_out_dir`.

Due to only the rank 0 process calling `setup_logger()`, it is only the
rank 0 process which modifies its global variable
`httomo.globals.run_out_dir`, and all other MPI processes do not modify
their copy of this global variable.

This in turn eventually causes a failure in saving any intermediate
dataset (including the final output of the pipeline). This is because
`httomo.globals.run_out_dir` is used in naming the output filepath, and
when it has its initial value of `None` the file understandably cannot
be saved, due to the output filepath string having `None` within the
directory portion of the path.

This change fixes this issue by moving the redefinition of
`httomo.globals.run_out_dir` outside of the `setup_logger()` function,
and thus changing it from the initial value of `None` in all MPI
processes, not just in the rank 0 process.