Add manually specified packages needed for tests during build
It is understood that automatic discovery of packages is preferred, as
in
https://setuptools.pypa.io/en/latest/userguide/package_discovery.html#automatic-discovery.
However, the package `httomo.data.hdf._utils` was faiing to be
automatically discovered (most likely due to the leading underscore in
`_utils`). So, for now, the packages necessary for the import tests in
the build process to pass have been manually listed.