Copy YAML files needed for tests into test env
As described in this documentation page for `conda build`
https://docs.conda.io/projects/conda-build/en/latest/resources/define-metadata.html#test-files,
the `test.files` section in `meta.yaml` can be used to provide any files
necessary for tests to run when the build process gets to the test
phase. Note that the files must be in the recipe directory for this
file-copying to work.

However, the YAML files necessary for the tests to run are not located
in the recipe directory. Instead, they are in the directories:
- `samples/`
- `templates/`

To avoid moving both the `samples/` and `templates/` directory to the
`conda/recipe/` directory, symlinks have been placed inside the
`conda/recipe/` directory to point to these directories.

Therefore, when `conda build` looks for the YAML files to copy into the
test environment, it follows these symlinks to the top-level `samples/`
and `templates/` directories and copies the correct files to the test
environment.