Define package version explicitly in `pyproject.toml`
Also, fetch package metadata information from `pyproject.toml` and reuse
in `meta.yaml` file.

It's worth noting that this change was mainly done to fix a strange bug
where the CLI version test `test_cli_version_shows_version()` was
unexpectedly failing during the test phase of the build process.

More specifically, the issue was that the value of `__version__` in the
test was correctly reported as `"0.0.1"`, but the result of the
`subprocess` call was seemingly reporting an incorrect value of
`"0.0.0"`.

After some investigation, it was discovered that `conda list` and `pip
list` reported different versions for the `httomo` package in the test
environment (`0.0.0` and `0.0.1` respectively). It was hypothesised that
this could be related to the mismatched version numbers that was only
present during the build process (the issue was not present in either
running the tests from the test environment leftover from the failed
build, nor from a fresh test environment created via `conda debug`).

After inspecting the `pyproject.toml`, it appeared to be configured to
dynamically fetch the package version, but no `tool.setuptools.dynamic`
field was defined to specify where it should be fetching it from.

There is some logic in `httomo/__init__.py` to determine the package
version which is the most likely place to point
`tool.setuptools.dynamic` to for the package version. Note however that
this logic was using `importlib.metadata.version()` to fetch the package
version from the installed package's metadata.

From this, it was unclear if `tool.setuptools.dynamic` could be pointed
to `__version__ ` defined in `httomo/__init__.py` to get the package
version. This is because it was unclear where this process would
actually fetch the version number from, as on first glance it appears to
be a sort-of infinite loop (ie, at buildtime the `pyproject.toml` file
tries to fetch the package version from `httomo/__init__.py`, which then
tries to get it from the package metadata in `pyproject.toml`, and so
on?).

Therefore, for now with this change, the single source of truth for the
package version will be hardcoded in `pyproject.toml`, and the
`meta.yaml` file will fetch the package version accordingly.

One last note is that the `pyproject.toml` file has some configuration
to fetch the version number from git tags, but which is currently not in
use. So in the future, it may well be that the package version number is
once again "dynamically defined" in `pyproject.toml`, but via fetching
from a git tag.