Fix incorrect type-checking of `in_data` arg for `check` command
The type of `in_data` is specifically `PosixPath` on linux rather than
`Path`. This error in judgement resulted in the `check` command silently
failing to pass the `in_data` value along to the YAML checker function
even if it was provided to the CLI.

A test has also been added to check that the output confirms that the
input data has been passed through by the CLI correctly to the YAML
checker function.