Improve package and YAML file discovery in `pyproject.toml`
Previously, packages and subpackages were being manually listed in
`pyproject.toml` for installation, which is a somewhat error-prone
method and thus undesirable if a better alternative exists.

In particular, there were several directories that had been added in
previous commits but were not being copied during installation, due to
them not being added to this list, and also not being given an
`__init__.py` file to mark them as a package.

Additionally, YAML files in these newly added directories were not being
captured by the globbing in the `[tool.setuptools.packages.find]`
section, and thus were also not being copied during installation.

The package discovery has been updated to find any subpackages in the
`httomo/` directory automatically, rather than adding each one in
manually.

The YAML file discovery has been updated to find all the necessary YAML
files in the `httomo/methods_database/` directory and subdirectories.
The `MANIFEST.in` file has been removed as it doesn't appear to affect
the YAML file discovery with the changes to the
`[tool.setuptools.packages.find]` section.

These changes also fix the failing tests that existed in the building of
the conda package caused by the incorrect installation (but were not
being detected, due to not having any CI setup for it yet).