Modify `BaseWrapper` class to work with only NumPy installed
There are a few cases where the use of `xp` to be a generic reference to
either NumPy or CuPy doesn't quite work (such as when `xp` is assumed to
have functions such as `asnumpy()` which only exist in CuPy and not
NumPy).

The minimal amount of code has been changed to enable all the pipeline
tests in `test_pipeline.py` to pass, though it's possible that more
modifications to the use of `xp` will be necessary if more CPU pipeline
tests with different CPU methods are added.