Use CPU median filter for multi-input pipeline
This pipeline file is only used in the `test_multi_inputs_pipeline`
test. The main aim of this test is to verify that the multi-input
functionality of httomo behaves as expected, so it's unnecessary to
require that a GPU be available to run this test, a CPU method will
suffice to carry out the intent of the test.