Run only 2 processes in param sweep+MPI test
Requesting 4 processes causes the
`test_sweep_pipeline_with_save_all_using_mpi` test to fail due to 4
"slots" not being available on the machine that the CI workflow runs on.
This test is only needing to compare the behaviour of parameter sweeps
between serial and parallel/MPI runs, so as long as there is more than 1
process being launched for the MPI run, it doesn't matter if it's 2 or 4
processes.