Filter out `ImportWarning` when running tests
This is primarily as a temporary fix to the following warning from
tomopy appearing during CI running CPU tests and causing the tests to
fail:
```
ImportWarning: TomoPy links to compiled components which are installed
separately and loaded using ctypes.util.find_library().Some
functionality is unavailable because an optional shared library,
tomo-accel, is missing.
```

Perhaps in the future it would be possible to address this issue more
directly by understanding why this optional shared library is
unavailable when tomopy is installed into the conda environment in the
CI run.