Move reslicing to the sections loop
The main changes are:
- the reslice operation now occurs after every section has completed
  (apart from the last section)
- `reslice_bool_list` has been removed from the `ResliceInfo` dataclass,
  because reslicing is now handled at the section-level rather than at
  the method-level
- the `_check_if_should_reslice()` function has been removed because
  reslicing is now determined by when a section changes

Note that this change has now moved part of the logic needed to run a
pipeline (the reslicing logic) to the sections loop, but it has not
moved other logic necessary for running the pipeline (such as actually
running the methods) into the sections loop.

This means that there is different logic necessary to run a pipeline now
located in two different areas: reslicing is in the "new area" (the
sections loop) whereas all other logic is still in the "old area" (in
the `run_method()` function). Therefore, this change by itself has
understandably broken all tests needing to run a pipeline.