Remove `reslice_ahead` and `should_reslice`
The main changes:
- `reslice_ahead` has been removed from `MethodFunc` because reslicing
  is handled at the section-level rather than the method-level
- `should_reslice` has been removed for the same reason as
  `reslice_ahead`
- `reslice_info` is no longer passed to `run_method()` because it no
  longer needs to be modified by `run_method()`
- remove the now unused `current_slice_dim` and `next_slice_dim`
  attributes on the `RunMethodInfo` dataclass

Note that although the tests are all passing locally, this does not mean
that the execution of a pipeline is working as intended. What's
happening now is that all reslices are occuring in the sections loop
(before any methods are being run), and then the methods are being run
without reslices inbetween them where appropriate. So, while these
changes enable the tests to pass locally, this behaviour is obviously
incorrect!