Create numpy array for output of sections containing recon method
Reconstruction methods change the shape of the input data (sinograms are
converted into slices of the volume), and so it is not straightforward
to overwrite regions of input data to a reconstruction method with
output of a reconstruction method due to mismatching shapes.

This in turn makes it difficult to overwrite the input data of a section
containing a reconstruction method with the output of that section,
since the reconstruction method within the section will have changed the
shape of the data.

As a workaround, if a reconstruction method is detected in a section, a
new numpy array is created with the appropriate shape to hold the full
reconstructed volume. As blocks are processed in the blocks-loop,
regions in this new array are where the processed block data is stored,
rather than attempting to overwrite data in the input array to the
section.