Add loader param for ignoring projections
For serial runs of httomo this functionality is fairly simple to
implement. However, for multi-process runs with MPI this functionality
becomes a bit more complicated.

One reason is that the ignored projections will belong to specific
processes and not to others, so there needs to be some logic for a
process to determine if it has any of the projections that have been
specified to be ignored.

Another reason is that, when it comes to saving the data to an hdf5
file, the `chunk.save_data_parallel()` function assumes that the full
data is evenly split along the slicing dimension and distributed among
the MPI processes. This is true when the initial read of the data has
been performed. However, when some projections are removed, and the fact
that some processes will remove projections whereas other processes
won't, this can lead to situations where different processes have
a different number of projections. This resulted in some changes to
`chunk.save_data_parallel()` to handle the more general case when
projections can be unevenly distributed among the MPI processes.