Generalise `single_sino_reslice()` to uneven data shapes across MPI processes
Generally, the shape of the projection data that each MPI process has
need not be the same across all MPI processes. Prior to this change, the
gathering of a single sinogram to the rank 0 process only worked when
all processes had the same projection data shape, otherwise,`mpi4py`
would raise the error `MPI_ERR_TRUNCATE`.