Add class for reading pipeline YAML files
Functionality relating to the parsing of the pipeline YAML file, and
creation of the relevant objects for pipeline execution, has been moved
into the `PipelineReader` class.

To deal with pre-processing methods, pipeline YAML files are now
suggested to be split into 3 "stages":
- loading
- pre-processing (categorised as methods to be run before normalisation)
- main processing

This separation is implemented in a pipeline YAML file via the YAML
document separator `---`.

Due to the nature of loading/pre-processing being quite different from
how the main processing is executed in the GPU loop, the methods in each
stage do not necessarily need to be represented by the same object.
Therefore, two more dataclasses have been introduced: `LoaderInfo` and
`PreProcessInfo`. This allows loader methods and methods in the
pre-processing stage to not be forced to generate a `MethodFunc` object
(which contains more data), and instead can be executed in the
loading/pre-processing stage with only the information that is required.

Mutation of the dictionary of parameters for a method is now much more
contained within the `PipelineReader` class, though some mutations still
occur in other places such as `prerun.prerun_method()`.