Separate "one method per module" YAML check from others
It makes sense to perform different YAML checks in isolation, to then be
able to test them in isolation from one another.

Prior to this change, testing of `check_one_method_per_module()` was
also running other checks like `sanity_check()`, when it would be better
to isolate the YAML checks to have one test be running only one check.