Pass det x/y dims correctly in paganin memory hook test
The `paganin_filter_tomopy()` function from httomolibgpu expects the
axes for the full 3D data to have the order `(angles, det_y, det_x)`,
and therefore expects each image to have the axes ordered as
`(det_y, det_x)`.

However, the memory hook test for it was passing the image dimensions
(ie, the "non slicing dimensions") to the memory estimator as
`(det_x, det_y)`, which is incorrect.