Remove unnecessary out slice size in estimation
The 2D FFT and IFFT used in `paganin_filter_tomopy()` are both C2C, and
so in conjunction with `overwrite_x=True`, the results of both are able
to be written into their input arrays. Thus, no new array is created to
hold the result of either of these transforms.