Modify `FullFileDataSet.shape` to return global shape
`DataSetBlock.shape` returns the shape of the block it represents. It
does this by inheriting the `shape` property getter from `DataSet`,
whose implementation references `self._data.shape`. In an instance of
`DataSetBlock`, this is the shape of the block.

`DataSet.shape` returns the shape of the chunk it represents.

Thus, the `shape` property appears to be returning the shape of the
"full data" that is represented by the object, ie:
- `DataSetBlock` represents a block, so `DataSetBlock.shape` returns the
  shape of the block
- `DataSet` represents a chunk, so `DataSet.shape` returns the shape of
  the chunk

To be in accordance with this notion of returning the "full data that
the object represents", `FullFileDataSet.shape` should return the shape
of the global data that it represents, not a chunk.

(It seems like an insignificant change, but another reason for this is
that it plays a role in fixing the currently failing test
`test_dataset_store.py::test_reslice_single_block_multi_process[True]`).