Fix `DataSetBlock.chunk_shape` when created from a `FullFileDataSet`
Prior to this change, when a `DataSetBlock` object was created from a
`FullFileDataSet` object, the one-liner implementation for the
`chunk_shape` property getter on the `DataSetBlock` object returned the
global data shape held in the parent `FullFileDataSet` (due to
modifications in a previous commit).

When a `DataSetBlock` object is created from a `FullFileDataSet` object,
this change allows a `DataSetBlock` to report the correct chunk shape
(the shape of the chunk that the block was created from) via the
`chunk_shape` property.