Fix MPI test performing reslice via data store
The main change to this test has been to use a `FullFileDataSet` to hold
the global data, in the parametrised test case for using an hdf5 file as
the storage means underneath the data store.

This is because the start index for creating or reading a block is
always relative to a chunk. Put another way, the start index when
calling `DataSet.make_block()`, `FullFileDataSet.make_block()`, or
`DataSetStoreReader.read_block()` is always relative to the boundaries
of a particular chunk. And put yet another way, the start index of a
block has no information about an offset for the block with respect to
the global data indices, it's only an offset relative to a particular
chunk.

What was going wrong in the test prior to this change was a consequence
of using a `DataSet` object to hold the global data, from which
`DataSetBlock` objects are then created from.

In general, global data would not be held in a `DataSet` object, it is
an object used for holding a single chunk rather than global data. For
single process runs, this distinction doesn't matter much, as there is
only one process and one chunk (which is the global data). However, when
there is a multi-process run, it starts to matter whether the global
data is contained in a `DataSet` object or a `FullFileDataSet` object.

The reason is related to the fact that making blocks from a `DataSet`
object or a `FullFileDataSet` is done where the block start index is
relative to the chunk associated with the MPI process. For a `DataSet`
object, which is for representing a single chunk, any block start index
passed to `DataSet.make_block()` will be relative to the single chunk
inside the `DataSet` object. If the global data is contained in that
single `DataSet` object, then this is a situation where the chunk data
contained in a `DataSet` object is also the global data.

If two MPI processes use block start indices which are relative to a
chunk, then both processes will be:
- using the same block start indices to select their block
- while also referring to the same single chunk held inside the
  `DataSet` object

Therefore, both MPI processes will be selecting the same block from the
`DataSet` object. `DataSet` has no notion of "offsetting to select a
specific chunk" because `DataSet` is for containing a single chunk, not
for containing multiple chunks.

This is what was happening in the test prior to this change:
- both MPI processes had a copy of the global data in a `DataSet` object
- both MPI processes were using the same block start indices, which are
  assumed to be relative to a chunk
- because the `DataSet` object that both MPI processes had contained a
  single chunk (the global data), both MPI processes were selecting the
  same underlying numpy array from their respective `DataSet` object

and therefore, both MPI processes were actually selecting the same block
of projections from the numpy array in their respective `DataSet`
objects.

So, the block selecting process had been done incorrectly. In this test,
the block selecting process is also finding which blocks should be
written to the data store. This would cause any subsequent attempts to
read blocks from the store to result in unexpected data being retrieved
(because the blocks that were written to the store contained incorrect
data).

The assertions at the end of the test are checking that sinogram blocks
which were read from the store are as expected. The store had incorrect
data written to it, so the assertions were failing.

Using a `FullFileDataSet` object to hold the global data that starts the
test instead of a `DataSet` object allows the block-selection done by
the two different MPI processes to be offset by the chunk that each MPI
process is associated with. This results in the two MPI processes
retrieving the correct data from their respective `FullFileDataSet`
objects, and so these projection blocks are correctly written to the
store, unlike before.

To make the comparison easier between:
- the sinogram block that has been read from the store
- the original global data

a numpy array is now created at the start of the test which holds
the global data, and this numpy array is accessed to get the expected
sinogram block, rather than accessing the `FullFileDataSet` object.

Note that while the test passes in general, it has been observed that
sometimes it does not pass due to assertion errors. It appears that
there may be some sort of synchronisation issue, where the data in the
`DataSetBlock` instances that comes from the hdf5 file underneath
`FullFileDataSet` is sometimes seen as having zeros in certain places,
despite the block-write request to the store being made correctly (as
far as one can tell). This needs to be investigated further.