Use separate darks/flats loading function in loader
Note that the tests for the separate darks/flats loading function
`get_darks_flats()` cover all cases that the tests for the darks/flats
loading that was originally in `StandardTomoLoader` covered.

Ie, the tests passing for the separate `get_darks_flats()` function
confirm that all cases of darks/flats loading that were present in
`StandardTomoLoader` are covered in the newer function
`get_darks_flats()`.

Therefore, now that the darks/flats fetching logic in
`StandardTomoLoader` has been swapped out in favour of using the
separate function `get_darks_flats()`, all the darks/flats fetching
logic in `StandardTomoLoader` can safely be removed.