Pass previewed data to `FullFileDataSet` in loader
Instead of passing the full data from `StandardTomoLoader` into
`FullFileDataSet` + using a chunk index that is relative to the global
data, pass the previewed/cropped data from `StandardTomoLoader` into
`FullFileDataSet` + use a chunk index relative to the previewed data.

Tests that check the chunk index for different previewing configs are
now unnecessary, because a chunk index is always relative to the
previewed data, which results in zeros in non-slicing dims. For example,
for projection chunks, the chunk index in the detector y and detector x
dimensions will always be zero, and this case is already covered by a
test that has not been removed.