Pass full h5py `Dataset` from loader to `FullFileDataSet`
This is reverting the change that passed sliced data from the h5py
`Dataset` object to `FullFileDataSet`, because that caused a read of the
selected region of data, which is counter to what `StandardTomoLoader`
is intended to do (it should read individual blocks from the selected
region of data, not the entire selected region at once).

This also requires reintroducing the tests that check the chunk index of
the loader, because the chunk index is now non-trivial due to it
reflecting any previewing that has been applied to the data.