Add `make_block()` implementation to `FullFileDataSet`
The main reason for this is so that start/stop indices of a block read
request are shifted according to the global index (the index of the
chunk associated with the current process, relative to the global data)
when using `FullFileDataSet`. For example, when the backing of the data
store is an hdf5 file (which contains the global data), each process
needs to shift its indices when reading a block to make sure the read is
performed in the chunk associated with the process.

This change also moves some code common to both `DataSet.make_block()`
and `FullFileDataSet.make_block()` to the constructor of `DataSetBlock`.