Use `data_offset` rather than `global_index` in loader
This change moves some of the index shifting information required for
excluding darks/flats from `FullFileDataSet.global_index` to a new
attribute `FullFileDataSet.data_offset`. Only `StandardTomoLoader` makes
use of `FullFileDataSet.data_offset`, no other objects need it (other
objects should instead exclusively use `FullFileDataSet.global_index`
for index shifting when reading blocks).

This change was motivated by a bug in which the data's slicing
dimension's length was being modified during a block write in pipeline
execution, which is not allowed.

Note that this change should be considered a workaround rather than a
proper fix, because the use of `FullFileDataSet` in `StandardTomoLoader`
appears more and more like a suboptimal design as development
progresses.