Use block's shape for centering's middle sino index
The rotation wrapper assumes that the section it will run in is a
projection section. This guarantees that the shape of the block that the
rotation wrapper receives will have the full detector y dimension length
(up to previewing). This means that it's fine to rely on the detector y
length of the block shape, rather than using the base dataset's detector
y length, to calculate the middle sinogram index.