Create block object directly inside loader
Many block reading tests using the loader are now broken due to moving
away from using `FullFileDataSet` to generate blocks for the loader, but
some have been fixed in this change.

In particular, the block reading loader tests that have been fixed are
the ones that don't preview/crop the data to be loaded, and the ones
that still need to fixed are the ones that use previewing (and thus
require some offsetting to do the block reading).