Remove darks/flats from max slices estimation
A consequence of changes being done in parallel to how darks/flats are
handled in the pipelione mean that it'll soon no longer be necessary to
account for the darks/flats in the max slices estimation.

Doing the removal here makes it simpler to handle darks/flats from the
loader's perspective, because now that the max slices estimation doesn't
need the darks/flats, the task runner won't need to ask the loader for
the darks/flats. This means that the loader doesn't need the behaviour
of translating empty darks/flats in the form of `None` into empty arrays
(which is currently only in `DataSetBlock`).

Put another way, removing the darks/flats input to the max slices
calculation here allows the avoidance of copying the logic in
`DataSetBlock` to translate `None` darks/flats values to empty arrays,
because the task runner no longer needs dark/flats from the loader for
max slices calculation. (Putting the logic into `AuxiliaryData` to be
usable by both the loader and the block class was also an option, but
it felt messy to give the `AuxiliaryData` the slicing dimension and data
shape in order to produce the appropriately shaped empty array in the
detector y and detector x dimensions).

Now, only `DataSetBlock` has the need to be able to translate `None`
darks/flats values to appropriately shaped empty arrays.