Remove angles length consistency check against projection dim length
The original intention of this check was to protect against having less
angles than projections. However, in the case of a pipeline involving
360 data, the angles get truncated to half their original length, while
the reconstruction shape in the 0th dimension remains (in general) much
larger than the number of angles after truncation.

This check perhaps could be moved to the loader to check when angles are
being loaded. Alternatively, the numpy array containing the angles in
the `AuxiliaryData` could be locked/ not writeable by default to provide
a higher level of certainty that the angles are only modified in a few
exempt situations, as a protection measure against the the angles array
becoming inconsistent with the data.