Temporarily comment-out pipeline test relying on ignoring darks/flats functionality
The ability to ignore darks/flats hasn't yet been added to the new
loader (the one that can load individual blocks), so this test doesn't
provide anything useful at the moment (it passes while not actually
ignoring any darks/flats). It should be added back in when the required
functionality has been added to the loader.