Avoid `res` var keeping reference to output cupy array in next iteration
This change is to address the observation that, when running the GPU
normalisation method with no other methods in the pipeline, it appeared
that the cupy array holding the output of the normalisation method for
block `n` was not getting freed before the transfer of block `n+1` from
CPU to GPU, causing GPU memory to overflow and causing a CUDA OOM error.
For example, the cupy output array produced when processing of block 0
finished was not freed before the input numpy array for block 1 was
transferred to GPU memory as a cupy array, which is unintended
behaviour.

Prior to this change, the `res` variable increased the reference count
of the `DataSetBlock` instance that `block` referred to by 1. This meant
that in iteration `n`, the `DataSetBlock` instance containing the output
cupy array was referred to by both `block` and `res`, so it had a
reference count of 2 at this point.

When execution went back to the start of the loop for iteration `n+1`,
the `block` variable got reassigned to a new `DataSetBlock` instance.
This decremented the reference count of the `DataSetBlock` instance
holding the output cupy array from iteration `n` by 1, so that now had a
reference count of 1.

However, the `res` variable was still referring to the `DataSetBlock`
instance holding the output cupy array from iteration `n`, which means
that the output cupy array from iteration `n` wouldn't yet get cleaned
up by the garbage collector.

It wouldn't be until in iteration `n+1` when the `res` variable got
reassigned to a different `DataSetBlock` instance (to the one from
iteration `n+1` rather than the one from iteration `n`) that:
- the reference count of the `DataSetBlock` from iteration `n` would
  drop to 0
- and thus the reference count of the output cupy array from iteration
  `n` would drop to 0 (which would make it ready to be cleaned up by the
  garbage collector, which would free the GPU memory allocated for it)

This was too late for the output cupy array from iteration `n` to be
freed in terms of available GPU memory, because the reassignment of
`res` wouldn't happen until the transfer of the numpy array to GPU
memory for iteration `n+1` was complete (which is part of the right hand
side of the reassignment of `res`).

Thus, the output cupy array from iteration `n` and the input array for
iteration `n+1` was attempted to be held in GPU memory at the same time.
This is unintended behaviour, the max slices estimation isn't intended
to address this situation, so GPU memory understandably ran out under
these circumstances.

This change removes the unnecessary `res` variable, which avoids the
extra reference to the `DataSetBlock` instance for iteration `n`. This
avoids having any references to the `DataSetBlock` instance from
iteration `n` within iteration `n+1`. This in turn allows the output
cupy array from iteration `n` to be cleaned up by the garbage collector
before the numpy array for iteration `n+1` gets transferred to GPU
memory.